refactor: extract field to it's own file
